pyproject.toml
[tool.poetry]
name = "literature"
version = "0.1.0"
description = ""
authors = ["eliroc98 <elisabetta.rocchetti0@gmail.com>"]
readme = "README.md"
package-mode=false

[tool.poetry.dependencies]
python = ">=3.11,<3.15"
sentence-transformers = "^5.2.0"
faiss-cpu = "^1.13.2"
pymupdf = "^1.26.7"
numpy = "^2.4.1"
streamlit = "^1.53.1"
nltk = "^3.9.2"
requests = "^2.32.0"
grobid-client-python = "^0.1.4"
torch = "^2.10.0"
pyyaml = "^6.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
